+++ platform/cc/shaper/FontRunIterator.hh
@@ -12,7 +12,9 @@ public:
                     const char* requestName,
                     SkFontStyle requestStyle,
                     const SkShaper::LanguageRunIterator* lang,
-                    std::shared_ptr<UBreakIterator> graphemeIter)
+                    std::shared_ptr<UBreakIterator> graphemeIter,
+                    bool approximateSpaces,
+                    bool approximatePunctuation)
         : fCurrent(utf8)
         , fBegin(utf8)
         , fEnd(fCurrent + utf8Bytes)
@@ -24,13 +26,30 @@ public:
         , fRequestStyle(requestStyle)
         , fLanguage(lang)
         , fGraphemeIter(graphemeIter)
+        , fApproximateSpaces(approximateSpaces)
+        , fApproximatePunctuation(approximatePunctuation)
     {
         fFont.setTypeface(font.refTypefaceOrDefault());
         fFallbackFont.setTypeface(nullptr);
     }
 
-    FontRunIterator(const char* utf8, size_t utf8Bytes, const SkFont& font, sk_sp<SkFontMgr> fallbackMgr, std::shared_ptr<UBreakIterator> graphemeIter)
-        : FontRunIterator(utf8, utf8Bytes, font, std::move(fallbackMgr), nullptr, font.refTypefaceOrDefault()->fontStyle(), nullptr, graphemeIter) {
+    FontRunIterator(const char* utf8,
+                    size_t utf8Bytes,
+                    const SkFont& font,
+                    sk_sp<SkFontMgr> fallbackMgr,
+                    std::shared_ptr<UBreakIterator> graphemeIter,
+                    bool approximateSpaces,
+                    bool approximatePunctuation)
+        : FontRunIterator(utf8,
+                          utf8Bytes,
+                          font,
+                          std::move(fallbackMgr),
+                          nullptr,
+                          font.refTypefaceOrDefault()->fontStyle(),
+                          nullptr,
+                          graphemeIter,
+                          approximateSpaces,
+                          approximatePunctuation) {
     }
 
     ~FontRunIterator() {
@@ -62,4 +81,6 @@ private:
     SkFontStyle const fRequestStyle;
     SkShaper::LanguageRunIterator const * const fLanguage;
     std::shared_ptr<UBreakIterator> fGraphemeIter;
+    bool fApproximateSpaces;
+    bool fApproximatePunctuation;
 };