+++ script/clean.py
@@ -7,6 +7,11 @@ def main():
   shutil.rmtree('platform/build', ignore_errors = True)
   shutil.rmtree('platform/target', ignore_errors = True)
   shutil.rmtree('tests/target', ignore_errors = True)
+  shutil.rmtree('examples/lwjgl/target', ignore_errors = True)
+  shutil.rmtree('examples/kwinit/target', ignore_errors = True)
+  shutil.rmtree('examples/jwm/target', ignore_errors = True)
+  shutil.rmtree('examples/swt/target', ignore_errors = True)
+
   return 0
 
 if __name__ == '__main__':