+++ shared/java/shaper/Shaper.java
@@ -97,29 +97,25 @@ public class Shaper extends Managed {
 
     @Nullable @Contract("_, _ -> new")
     public TextBlob shape(String text, Font font) {
-        return shape(text, font, null, true, Float.POSITIVE_INFINITY, Point.ZERO);
+        return shape(text, font, ShapingOptions.DEFAULT, Float.POSITIVE_INFINITY, Point.ZERO);
     }
 
     @Nullable @Contract("_, _, _ -> new")
     public TextBlob shape(String text, Font font, float width) {
-        return shape(text, font, null, true, width, Point.ZERO);
+        return shape(text, font, ShapingOptions.DEFAULT, width, Point.ZERO);
     }
 
     @Nullable @Contract("_, _, _, _ -> new")
     public TextBlob shape(String text, Font font, float width, @NotNull Point offset) {
-        return shape(text, font, null, true, width, offset);
+        return shape(text, font, ShapingOptions.DEFAULT, width, offset);
     }
 
     @Nullable @Contract("_, _, _, _, _ -> new")
-    public TextBlob shape(String text, Font font, boolean leftToRight, float width, @NotNull Point offset) {
-        return shape(text, font, null, leftToRight, width, offset);
-    }
-
-    @Nullable @Contract("_, _, _, _, _, _ -> new")
-    public TextBlob shape(String text, Font font, @Nullable FontFeature[] features, boolean leftToRight, float width, @NotNull Point offset) {
+    public TextBlob shape(String text, Font font, @NotNull ShapingOptions opts, float width, @NotNull Point offset) {
         try {
+            assert opts != null : "Can’t Shaper::shape with opts == null";
             Stats.onNativeCall();
-            long ptr = _nShapeBlob(_ptr, text, Native.getPtr(font), features, leftToRight, width, offset._x, offset._y);
+            long ptr = _nShapeBlob(_ptr, text, Native.getPtr(font), opts, width, offset._x, offset._y);
             return 0 == ptr ? null : new TextBlob(ptr);
         } finally {
             Reference.reachabilityFence(this);
@@ -127,22 +123,20 @@ public class Shaper extends Managed {
         }
     }
 
-    @NotNull @Contract("_, _, _, _, _, _, _ -> this")
+    @NotNull @Contract("_, _, _, _, _ -> this")
     public Shaper shape(String text,
                         Font font,
-                        @Nullable FontMgr fontMgr,
-                        @Nullable FontFeature[] features,
-                        boolean leftToRight,
+                        @NotNull ShapingOptions opts,
                         float width,
                         RunHandler runHandler)
     {
         try (ManagedString textUtf8 = new ManagedString(text);
-             FontMgrRunIterator fontIter = new FontMgrRunIterator(textUtf8, false, font, fontMgr);
-             IcuBidiRunIterator bidiIter = new IcuBidiRunIterator(textUtf8, false, leftToRight ? java.text.Bidi.DIRECTION_LEFT_TO_RIGHT : java.text.Bidi.DIRECTION_RIGHT_TO_LEFT);
+             FontMgrRunIterator fontIter = new FontMgrRunIterator(textUtf8, false, font, opts);
+             IcuBidiRunIterator bidiIter = new IcuBidiRunIterator(textUtf8, false, opts._leftToRight ? java.text.Bidi.DIRECTION_LEFT_TO_RIGHT : java.text.Bidi.DIRECTION_RIGHT_TO_LEFT);
              HbIcuScriptRunIterator scriptIter = new HbIcuScriptRunIterator(textUtf8, false);)
         {
             Iterator<LanguageRun> langIter = new TrivialLanguageRunIterator(text, Locale.getDefault().toLanguageTag());
-            return shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, features, width, runHandler);
+            return shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, opts, width, runHandler);
         }
     }
 
@@ -152,12 +146,12 @@ public class Shaper extends Managed {
                         @NotNull Iterator<BidiRun> bidiIter,
                         @NotNull Iterator<ScriptRun> scriptIter,
                         @NotNull Iterator<LanguageRun> langIter,
-                        @Nullable FontFeature[] features,
+                        @NotNull ShapingOptions opts,
                         float width,
                         @NotNull RunHandler runHandler)
     {
         try (ManagedString textUtf8 = new ManagedString(text);) {
-            return shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, features, width, runHandler);
+            return shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, opts, width, runHandler);
         }
     }
 
@@ -167,7 +161,7 @@ public class Shaper extends Managed {
                         @NotNull Iterator<BidiRun> bidiIter,
                         @NotNull Iterator<ScriptRun> scriptIter,
                         @NotNull Iterator<LanguageRun> langIter,
-                        @Nullable FontFeature[] features,
+                        @NotNull ShapingOptions opts,
                         float width,
                         @NotNull RunHandler runHandler)
     {
@@ -175,20 +169,27 @@ public class Shaper extends Managed {
         assert bidiIter != null : "BidiRunIterator == null";
         assert scriptIter != null : "ScriptRunIterator == null";
         assert langIter != null : "LanguageRunIterator == null";
+        assert opts != null : "Can’t Shaper::shape with opts == null";
         Stats.onNativeCall();
-        _nShape(_ptr, Native.getPtr(textUtf8), fontIter, bidiIter, scriptIter, langIter, features, width, runHandler);
+        _nShape(_ptr, Native.getPtr(textUtf8), fontIter, bidiIter, scriptIter, langIter, opts, width, runHandler);
         return this;
     }
 
-    @NotNull @Contract("_, _, _, _, _, _ -> new")
-    public TextLine shapeLine(String text, Font font, @Nullable FontFeature[] features, boolean leftToRight) {
+    @NotNull @Contract("_, _, _ -> new")
+    public TextLine shapeLine(String text, Font font, @NotNull ShapingOptions opts) {
         try {
+            assert opts != null : "Can’t Shaper::shapeLine with opts == null";
             Stats.onNativeCall();
-            return new TextLine(_nShapeLine(_ptr, text, Native.getPtr(font), features, leftToRight));
+            return new TextLine(_nShapeLine(_ptr, text, Native.getPtr(font), opts));
         } finally {
             Reference.reachabilityFence(this);
             Reference.reachabilityFence(font);
         }
+    }
+
+    @NotNull @Contract("_, _, _ -> new")
+    public TextLine shapeLine(String text, Font font) {
+        return shapeLine(text, font, ShapingOptions.DEFAULT);
     }
 
     @ApiStatus.Internal
@@ -208,8 +209,7 @@ public class Shaper extends Managed {
     public static native long _nMakeShapeDontWrapOrReorder(long fontMgrPtr);
     public static native long _nMakeCoreText();
     public static native long _nMake(long fontMgrPtr);
-    public static native long _nShapeBlob(long ptr, String text, long fontPtr, FontFeature[] features, boolean leftToRight, float width, float offsetX, float offsetY);
-    public static native long _nShapeLine(long ptr, String text, long fontPtr, FontFeature[] features, boolean leftToRight);
-    public static native void _nShape(long ptr, long textPtr, Iterator<FontRun> fontIter, Iterator<BidiRun> bidiIter, Iterator<ScriptRun> scriptIter, Iterator<LanguageRun> langIter,
-                                      FontFeature[] features, float width, RunHandler runHandler);
+    public static native long _nShapeBlob(long ptr, String text, long fontPtr, ShapingOptions opts, float width, float offsetX, float offsetY);
+    public static native long _nShapeLine(long ptr, String text, long fontPtr, ShapingOptions opts);
+    public static native void _nShape(long ptr, long textPtr, Iterator<FontRun> fontIter, Iterator<BidiRun> bidiIter, Iterator<ScriptRun> scriptIter, Iterator<LanguageRun> langIter, ShapingOptions opts, float width, RunHandler runHandler);
 }